Add threaded downloads from Chef Servers
Since Chef Server stores cookbooks as individual files, many downloads
are needed to transfer a cookbook. It's faster to download in parallel.

This only works when the HTTP client passed into the location object is
properly multi-threaded, though, so by default the concurrency is set to
1. In a future update to ChefDK, we can add a multi-thread wrapper
around `Chef::ServerAPI` to enable this.

Signed-off-by: Daniel DeLeo <dan@chef.io>