Change cache validation to always check sums
Update cache validation to always check the checksums of individual
files instead of assuming that the content is correct if the directory
is present.   This will also remove any files that should not be in the
cache directory.

Combined, these changes give a better chance that uploaded artifact will
match the downloaded cookbook.

Reference
https://github.com/chef/chef-workstation/issues/1273#issuecomment-675566249
for details.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>